packages/casbin-fastapi-decorator-casdoor/pyproject.toml
[project]
name = "casbin-fastapi-decorator-casdoor"
version = "0.1.0"
description = "Casdoor OAuth2 user provider and enforcer for casbin-fastapi-decorator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "casbin-fastapi-decorator",
    "casdoor>=1.28.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["uv_build>=0.9.18,<0.11.0"]
build-backend = "uv_build"

[tool.uv.sources]
casbin-fastapi-decorator = { workspace = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"
python_files = ["test_*.py", "*_test.py", "act.py", "error.py"]
markers = [
    "integration: integration tests (real I/O, HTTP)",
    "unit: unit tests (no real I/O)",
    "casdoor_provider: CasdoorUserProvider component tests",
    "casdoor_enforcer: CasdoorEnforcerProvider component tests",
]

[tool.bandit]
exclude_dirs = ["tests"]

[dependency-groups]
test = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "httpx>=0.28.0",
]
